Refactor files to functionality/CreateSchedule directory
