+++ frontend/src/functionality/CreateSchedule/CreateSchedule.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import TimeBlockForm from '../../components/TimeBlockForm.jsx';
-import GeneratorForm from '../../components/GeneratorForm.jsx';
+import GeneratorForm from './GeneratorForm.jsx';
 import { useCreateTimeBlock } from './utils/Hooks/useCreateTimeBlock.js';
 import { useGenerateSchedule } from './utils/Hooks/useGenerateSchedule.js';
