+++ frontend/src/functionality/CreateSchedule/GlobalOptions.jsx
