+++ frontend/src/functionality/CreateSchedule/__tests__/GeneratorFormItem.test.jsx
