+++ frontend/src/functionality/CreateSchedule/__tests__/GlobalOptions.test.jsx
